requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
scipy==1.11.1

# GIS and Spatial Analysis
geopandas==0.13.2
shapely==2.0.1
rasterio==1.3.8
folium==0.14.0
pyproj==3.6.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework and API
fastapi==0.101.1
uvicorn==0.23.2
streamlit==1.25.0
pydantic==2.1.1

# Data Processing
pyyaml==6.0.1
requests==2.31.0
openpyxl==3.1.2

# Jupyter and Development
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0
loguru==0.7.0